# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AnyRouter 签到中台管理系统",
        "title": "AnyRouter 管理后台 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/config/telegram/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "发送 Telegram 测试消息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/config/{category}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "获取指定分类的配置",
                "parameters": [
                    {
                        "enum": [
                            "telegram",
                            "system"
                        ],
                        "type": "string",
                        "description": "配置分类",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "更新指定分类的配置",
                "parameters": [
                    {
                        "enum": [
                            "telegram",
                            "system"
                        ],
                        "type": "string",
                        "description": "配置分类",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置键值对",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/cron": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "获取所有定时任务",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/anyrouter-checkin_internal_model.CronTask"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "创建定时任务",
                "parameters": [
                    {
                        "description": "任务参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CronRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/anyrouter-checkin_internal_model.CronTask"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/cron/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "更新定时任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "任务参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CronRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/anyrouter-checkin_internal_model.CronTask"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "删除定时任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/cron/{id}/trigger": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "立即触发执行定时任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志"
                ],
                "summary": "获取签到日志列表与今日账号统计",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/anyrouter-checkin_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/anyrouter-checkin_internal_service.CheckinLogSummary"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "anyrouter-checkin_internal_model.CheckinLog": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/carbon.DateTime"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "anyrouter-checkin_internal_model.CronTask": {
            "type": "object",
            "properties": {
                "account_ids": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/carbon.DateTime"
                },
                "cron_expr": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_run": {
                    "$ref": "#/definitions/carbon.DateTime"
                },
                "name": {
                    "type": "string"
                },
                "next_run": {
                    "$ref": "#/definitions/carbon.DateTime"
                },
                "status": {
                    "type": "integer"
                },
                "task_type": {
                    "type": "string"
                }
            }
        },
        "anyrouter-checkin_internal_service.CheckinLogSummary": {
            "type": "object",
            "properties": {
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/anyrouter-checkin_internal_model.CheckinLog"
                    }
                },
                "today_checkin_account_count": {
                    "type": "integer"
                }
            }
        },
...
        "anyrouter-checkin_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "carbon.DateTime": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
        "internal_handler.CronRequest": {
            "type": "object",
            "required": [
                "cron_expr",
                "name"
            ],
            "properties": {
                "account_ids": {
                    "type": "string",
                    "example": "[1,2]"
                },
                "cron_expr": {
                    "type": "string",
                    "example": "0 8 * * *"
                },
                "name": {
                    "type": "string",
                    "example": "每日签到"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "task_type": {
                    "type": "string",
                    "example": "checkin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}